Fixes application of javac options (#18)
This PR fixes the syntax to enable `javacOptions` in each of the `sbt` modules.

This fix revealed a deprecation warning in akka-grpc generated code. The consequence of this warning with the `-Werror` flag is the build of this PR will break. Unfortunately the newet `0.5.0` release of `akka-grpc` introduces other compilation errors on the generated code so at the moment we can't have a 100% error-free lagom-grpc sample.